requirements.txt
anthropic
markdown
pygments
python-dotenv
cryptography
aiohttp>=3.8.0
pillow>=9.0.0